requirements.txt
tensorflow-cpu
nltk
colorama
flask
scikit-learn
keras
gunicorn
numpy
flask_cors